Set `awaitInitial...` to `true` by default, fix this config.
Setting to `false` requires additional care by the developer
because the data won't be available in the [State] right away,
so this "asynchronous" loading mode has been made as an opt-in
variant if the user is consciously chooses to use this mode
for performance optimization.

The functionality `awaitInitial...` == `true` was fixed to work
(it had a deadlock due to synchronous waiting for `Dispatchers.Main`).

Added kotlinx-coroutines-swing, which is required for
`Dispatchers.MAIN` to work.